Changed folder structure and separated css from js file
