+++ package.json
@@ -11,7 +11,7 @@
     "expo": "^30.0.1",
     "raven-js": "^3.27.0",
     "react": "16.5.2",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-28.0.0.tar.gz",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
     "react-native-app-intro": "^1.1.5",
     "react-native-app-intro-slider": "^1.0.1",
     "react-native-elements": "^0.19.1",
