+++ src/components/common/ListItems/Styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  todoItem: {
+    alignItems: 'center',
+    padding: 8,
+    width: 320,
+    borderBottomWidth: 1.5,
+    borderColor: '#e0e0e0',
+    backgroundColor: '#fff',
+    flex: 1,
+    flexDirection: 'row',
+  },
+  todoText: {
+    flex: 1,
+  },
+});
+
+export default styles;
