+++ src/components/common/Logo/Logo.js
@@ -1,33 +1,18 @@
 import React from 'react';
-import { View, Text, Image, StyleSheet } from 'react-native';
+import { View, Text, Image } from 'react-native';
+import styles from './Styles';
 
 const Logo = () => (
-	<View style={styles.imageOuterView}>
-		<Image
-			source={{
-				uri:
-					'https://d1qb2nb5cznatu.cloudfront.net/startups/i/4784928-963c639863cf77f25d3ec213c37db74c-medium_jpg.jpg?buster=1522462849'
-			}}
-			style={styles.imageStyle}
-		/>
-		<Text style={styles.titleText}>Groceristar</Text>
-	</View>
+  <View style={styles.imageOuterView}>
+    <Image
+      source={{
+        uri:
+          'https://d1qb2nb5cznatu.cloudfront.net/startups/i/4784928-963c639863cf77f25d3ec213c37db74c-medium_jpg.jpg?buster=1522462849',
+      }}
+      style={styles.imageStyle}
+    />
+    <Text style={styles.titleText}>Groceristar</Text>
+  </View>
 );
-
-const styles = StyleSheet.create({
-	imageOuterView: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center'
-	},
-	imageStyle: {
-		width: 100,
-		height: 100
-	},
-	titleText: {
-		fontWeight: '500',
-		fontSize: 30
-	}
-});
 
 export default Logo;
