+++ src/components/common/Logo/Styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  imageOuterView: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  imageStyle: {
+    width: 100,
+    height: 100,
+  },
+  titleText: {
+    fontWeight: '500',
+    fontSize: 30,
+  },
+});
+
+export default styles;
