+++ src/components/common/TextField/Styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  textStyle: {
+    color: '#000',
+  },
+  fonstSize: { fontSize: 17 },
+  inputContainer: {
+    width: '100%',
+    height: 47,
+    borderRadius: 3,
+    borderWidth: 1.5,
+    justifyContent: 'center',
+    paddingLeft: 10,
+    paddingRight: 10,
+  },
+});
+
+export default styles;
