+++ src/components/common/TextField/TextField.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, Text, TextInput, StyleSheet } from 'react-native';
+import { View, Text, TextInput } from 'react-native';
+import styles from './Styles';
 
 const TextField = ({ text }) => (
   <View>
@@ -9,21 +10,5 @@ const TextField = ({ text }) => (
     </View>
   </View>
 );
-
-const styles = StyleSheet.create({
-  textStyle: {
-    color: '#000'
-  },
-  fonstSize: { fontSize: 17 },
-  inputContainer: {
-    width: '100%',
-    height: 47,
-    borderRadius: 3,
-    borderWidth: 1.5,
-    justifyContent: 'center',
-    paddingLeft: 10,
-    paddingRight: 10
-  }
-});
 
 export default TextField;
