+++ src/components/layouts/AddValueToList/Styles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  inputContainer: { flex: 0.8 },
+  buttonContainer: { flex: 0.2 },
+  textInput: {
+    marginLeft: 5,
+    height: 35,
+    borderColor: 'gray',
+    borderWidth: 2,
+    padding: 4,
+  },
+  iconContainer: { padding: 4 },
+  addIngredients: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: 50,
+  },
+});
+
+export default styles;
