+++ src/components/layouts/GroceryList/Styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'rgba(242, 175, 82,0.8)',
+    paddingTop: 15,
+    paddingLeft: 30,
+  },
+});
+
+export default styles;
