+++ src/components/layouts/IngredientDetails/IngredientDetails.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { View, StyleSheet, FlatList, Text } from 'react-native';
+import { View, FlatList, Text } from 'react-native';
 import { CheckBox, Icon } from 'react-native-elements';
 import uuid from 'uuidv4';
 import underscore from 'underscore';
-
-import { getAllIngredientsList } from '../../HelperFunctions';
-import AddValueToList from './AddValueToList';
+import styles from './Styles';
+import { getAllIngredientsList } from '../../../HelperFunctions';
+import AddValueToList from '../AddValueToList/AddValueToList';
 
 let tempName = '';
 
@@ -163,31 +163,3 @@ export default class IngredientDetails extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'rgba(242, 175, 82,0.8)',
-    paddingTop: 15,
-    paddingLeft: 30,
-  },
-  addOldIngrdients: {
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'center',
-    margin: 10,
-    borderWidth: 2,
-    borderColor: '#657287',
-  },
-  ingredientName: {
-    flex: 0.9,
-    marginLeft: 10,
-    borderRightWidth: 2,
-    borderRightColor: '#000',
-  },
-  iconContainer: {
-    flex: 0.1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
