+++ src/components/layouts/Login/Styles.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: '#fff',
+    flex: 1,
+    width: '100%',
+    padding: 25,
+    justifyContent: 'center',
+  },
+  buttonOuterView: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  link: {
+    textDecorationLine: 'underline',
+    color: '#4985FF',
+    alignSelf: 'center',
+    marginTop: 5,
+  },
+});
+
+export default styles;
