+++ src/components/layouts/Todo/Styles.js
@@ -1,0 +1,27 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '#fff',
+    paddingTop: 25,
+  },
+  formView: {
+    borderBottomWidth: 1,
+    borderColor: '#ccc',
+    paddingBottom: 10,
+    paddingTop: 15,
+  },
+  inputForm: {
+    backgroundColor: '#fff',
+    width: 320,
+    height: 40,
+    padding: 8,
+    marginBottom: 10,
+    marginTop: 10,
+  },
+});
+
+export default styles;
